next batch

+++ src/Conversion/ONNXToKrnl/ControlFlow/Scan.cpp
@@ -132,7 +132,7 @@ struct ONNXScanOpLowering : public ConversionPattern {
       // thus becomes redundant.
       SmallVector<Value, 4> bodyOutputs(
           resultsRange.begin(), resultsRange.end());
-      for (int i = 0; i < bodyOutputs.size(); i++) {
+      for (unsigned int i = 0; i < bodyOutputs.size(); i++) {
         auto output = bodyOutputs[i];
         assert((output.getType().isa<TensorType>() ||
                    output.getType().isa<MemRefType>()) &&
@@ -314,7 +314,7 @@ struct ONNXScanOpLowering : public ConversionPattern {
     auto srcTy = src.getType().cast<MemRefType>();
     SmallVector<Value, 4> readIV(readPrefix.begin(), readPrefix.end());
     SmallVector<Value, 4> writeIV;
-    if (srcTy.getRank() > readIV.size()) {
+    if ((size_t) srcTy.getRank() > readIV.size()) {
       BuildKrnlLoop loop(rewriter, loc, srcTy.getRank() - readPrefix.size());
       loop.createDefineOp();
       for (int i = readIV.size(); i < srcTy.getRank(); i++)
